refactor : make class Racing in charge of Playing racingGame
Application 에 있던 게임 동작 메서드들을 클래스로 분리. 생성자 없이 호출하고 싶어서 내부 필드와 메서드들을 static으로 구현
